split off algorithmic part of training collection, for use in openEO udf
